+++ Models/OrderDetail.cs
@@ -17,9 +17,9 @@ namespace WebsiteBanCaPhe.Models
         public Product Product { get; set; }
 
         [Required]
-        public int Quantity { get; set; }
+        public long Quantity { get; set; }
 
-		[BindNever]
-		public decimal TotalPrice { get; set; }
+        [BindNever]
+        public long TotalPrice { get; set; }
     }
 }
